# Many various optimizations