These changes remove many repetitions of low level operations in frequently executed parts of the code, for significant performance gains. Measured performance gains include 10% update speed, 30% dequeue speed, and 28% remove speed on the fast queue, and 8.5% update speed, 25% contains speed, 9% dequeue speed, 25% remove speed, and smaller improvement in enqueue speed on the simple queue. Fast queue optimizations are also copied into the stable and generic queues for similar gains. Also, resize is nearly 4 times as fast by using a built-in array copy instead of a manual loop.

I didn't see any performance tests in the project, so I made my own for local testing as follows:
Fast queue, set size to 1000000, and create and store exactly that many basic nodes with just a string value.
- Enqueue all nodes, then clear queue, to make sure it's fully warmed up
- Enqueue all nodes with random priorities, and time it
- Update all nodes with new random priorities, and time it
- Check contains on all nodes, and time it
- Resize the queue to double, then back, 10 times each, and time it
- Dequeue until empty, and time it
- Re-enqueue all nodes, then remove all nodes, and time the remove part
- Repeat 20 times and average the times

Simple queue, create and store 1000000 `new object()` data items
- Enqueue all nodes, then clear queue, to warm up
- Enqueue all nodes with random priorities, and time it
- Update all nodes with new random priorities, and time it
- Check contains on all nodes, and time it
- Dequeue until empty, and time it
- EnqueueWithoutDuplicates all nodes, and time it
- Remove all nodes, and time it
- Repeat 20 times and average the times

I made no attempt to account for the overhead cost of the test loops and random number generation.

Result times in milliseconds:
Fast queue:

| Operation | Time Before | Time After |
| ----------- | ----- | ------- |
| Enqueue | 50 | 51 |
| Update | 96 | 87 |
| Contains | 6 | 6 |
| Resize | 409 | 106 |
| Dequeue | 992 | 762 |
| Remove | 147 | 115 |

Simple queue:

| Operation | Time Before | Time After |
| ----------- | ----- | ------- |
| Enqueue | 1080 | 1063 |
| Update | 419 | 386 |
| Contains | 216 | 172 |
| Dequeue | 1575 | 1442 |
| EnqueueWithoutDuplicates | 1143 | 1064 |
| Remove | 597 | 478 |

The decrease in fast queue enqueue speed is within random variation for the many times I ran these tests while making these changes. Fast queue contains is unaffected, all other operations are significantly faster.

By far the largest single improvement in the fast queue, besides the trivial resize change, was the rewrite of CascadeDown. Reducing the number of operations in Remove (and in the copy of it I put in Dequeue so I could remove another comparison or two) was a strong runner up, and several other changes had small but noticeable effects.

The simple queue benefited from the same changes through its wrapped GenericPriorityQueue, and also was significantly improved by various rewrites that reduce the number of times an object's hash is calculated and looked up in a dictionary, and dropping the calculation of pointless intermediate results in several places.

A further few percent improvement in most fast queue operations could be achieved by changing the priority type from `float` to `int`, or adding a copy of the fast queue with that difference - the generic queue won't do for this because the overhead of using a Comparer is too large even if the stability is removed - but I restricted myself to changes not visible to users.