@BlueRaja 2017-06-05
Awesome, great work!

It will take me a while to review this because I will need to see which changes actually make a difference in the JIT output.  For example, I highly doubt `nodeIndex / 2` to `nodeIndex >> 1` changes anything.

How did you get a 25% improvement to `Contains`?  I can't imagine doing anything simpler, and it seems like you haven't changed it at all.