@dougmill 2017-06-05
I was recently surprised in another project to learn that the return value can differ between `someInt / 2` and `someInt >> 1` - they round negative numbers differently. So, I don't know for certain that it will make a difference, but it seems plausible.

On contains, that improvement is only in the simple queue, I didn't see any possible improvement in the fast queue version.

For simple queue contains, my change drops the following operations:
When item is null, picking `_nullNodesCache[0]` or `null` and then comparing it to `null`.
In other cases, fetching the list of nodes (rather than just checking one exists), checking its count, picking `nodes[0]` or `null` based on the count, and comparing the result to `null`.

Skipping the count check is safe because any time a list is reduced to empty it is also removed from the dictionary, so there will never be an empty list in there. And apparently I neglected to apply that in `GetExistingNode`, so the final line of that should be changed to just `return nodes[0];` in addition to the changes already here.