@BlueRaja 2017-06-05
Oh I see, I should have looked at the results closer.  Yes I didn't spend a lot of time optimizing the SimpleQueue.  I'll gladly merge these changes once I've vetted them.  Please give me a week or two, I'm extremely busy recently.  Thanks!