@dougmill 2017-06-05
While you're at it, please add a version of fast queue with `int` type priority - the project I'm using this for produces `int` priority values naturally, and integer comparisons are faster than floating point. I tried it out without committing the change, and measured a low to mid single digit percent improvement, I don't remember exactly how much, in everything but Contains.