@BlueRaja 2017-06-11
I was seeing slowdowns in my pathfinding benchmarks, so I've added tests to benchmark individual Priority Queue methods (pull from master to see them).

`Enqueue` seems to be ~17% faster all around, but `UpdatePriority` is ~6% slower on large datasets, and `Dequeue` seems to be an order of magnitude slower (!?!).  There is the possibility I screwed something up...