@dougmill 2017-06-11
I would guess you screwed up in the CascadeDown rewrite somehow. That would not affect enqueue, would be partially balanced by CascadeUp improvement in update and also be relatively small because random nodes are likely near the bottom, and have a very large effect on dequeue with nothing to balance it and always a long path to swap through.