@BlueRaja 2017-06-11
Try uninstalling and reinstalling BenchmarkDotNet through NuGet