@dougmill 2017-06-11
Got it running, and duplicated your result. I think it has something to do with the `IterationSetup` bug you reported. I switched that method to be its own benchmark to give a baseline and then had all the others actually call it themselves, and dequeue became faster with my code.

Interestingly, the "just do the setup" benchmark is slower with my changes than without. I'm not entirely sure why. I thought it might be my change adding a little up front overhead that never gets compensated for because you're enqueueing in strict ascending priority order so it always bails out immediately, but reverting that didn't meaningfully help.

I might try applying one change at a time, as small as I can split them up, and running the benchmarks at each step, but honestly I'm really not that confident in these numbers in the first place. Something very strange is clearly going on because I don't see how my changes could _**possibly**_ affect the behavior of that bug, and it seems such a basic thing that they really shouldn't have missed it. So, this benchmarking tool spits out a lot of numbers and bits of analysis, but at this point I don't trust those numbers to be correct.