@BlueRaja 2017-06-11
Well that's annoying, I assumed that such a popular benchmarking library would actually work correctly.  I'll try rewriting the benchmarks to combine enqueuing + dequeuing tomorrow.