@dougmill 2017-06-11
Hmm, per the comments on your issue report, it's apparently a design issue with "Iteration" meaning something a bit higher level than one method invocation. I'll be doing some experimenting later.

I've tweaked your test cases a bit, because I think always doing things with `Queue.First` and never an arbitrary node, and always using strictly in order priorities, probably skew the behavior patterns. I'm currently running with this:
```
namespace Priority_Queue_Benchmarks
{
    public class Benchmarks
    {
        [Params(1, 100, 10000)]
        public int QueueSize;

        private FastPriorityQueue<FastPriorityQueueNode> Queue;
        private FastPriorityQueueNode[] Nodes;
        private int[] RandomPriorities;

        [GlobalSetup]
        public void GlobalSetup()
        {
            Queue = new FastPriorityQueue<FastPriorityQueueNode>(QueueSize * 2);
            Nodes = new FastPriorityQueueNode[QueueSize];
            Random rand = new Random();
            RandomPriorities = new int[QueueSize * 2];
            for (int i = 0; i < QueueSize * 2; i++)
            {
                RandomPriorities[i] = rand.Next(16777216); // constrain to range float can hold with no rounding
            }
        }

        [Benchmark]
        public void IterationSetup()
        {
            Queue.Clear();
            for (int i = 0; i < QueueSize; i++)
            {
                Nodes[i] = new FastPriorityQueueNode();
                Queue.Enqueue(Nodes[i], RandomPriorities[i + QueueSize]);
            }
        }

        [Benchmark]
        public void Dequeue()
        {
            IterationSetup();
            for (int i = 0; i <= QueueSize / 2; i++)
            {
                Queue.Dequeue();
            }
        }

        [Benchmark]
        public void Enqueue()
        {
            IterationSetup();
            for (int i = 0; i <= QueueSize / 2; i++)
            {
                Queue.Enqueue(new FastPriorityQueueNode(), RandomPriorities[i]);
            }
        }

        [Benchmark]
        public bool Contains()
        {
            IterationSetup();
            bool ret = true;
            for (int i = 0; i < QueueSize; i++)
            {
                ret &= Queue.Contains(Nodes[i]);
            }
            return ret;
        }

        [Benchmark]
        public void Clear()
        {
            IterationSetup();
            Queue.Clear();
        }

        [Benchmark]
        public void UpdatePriority()
        {
            IterationSetup();
            for (int i = 0; i < QueueSize; i++)
            {
                Queue.UpdatePriority(Nodes[i], RandomPriorities[i]);
            }
        }
    }
}
```